Check the result of the optimization
As programmed, the optimization does nothing. This is a bug I
encountered during my first attempt to code this method, back in Fall
2023.

The fix I found back then is very unsatisfactory. Namely it is to force
the first element of the optimal filter's denominator coefficients to
1, as opposed to leave it as an optimization variable.

I will try this again in the next commit.