Fix documentation
Prior to this commit, but as of the very previous one, the documentation
did not accurately reflect the form imposed on the filter the method is
looking for. The leading coefficient of the denominator polynomial is
fixed at 1e0. This is a crucial difference that fixes a bug recorded in
another commit.